requirements.txt
Flask==1.1.2
gunicorn==19.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
keras==2.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
monotonic==1.5
mysql-client==0.0.1
mysql-connector-python==8.0.27
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
wadllib==1.3.4
Werkzeug==2.0.1

